Move Modals next to where they are used
